Refactor Header and Footer components for TripCraft AI. Update import paths to use absolute imports and create new Header and Footer components for improved layout and organization.
