+++ submissions/TripCraftAI/client/app/page.tsx
@@ -7,8 +7,8 @@ import {
   Calendar,
   Sparkles,
 } from "lucide-react";
-import Header from "./components/Header";
-import Footer from "./components/Footer";
+import Header from "@/components/header";
+import Footer from "@/components/footer";
 import { Button } from "@/components/ui/button";
 import {
   Card,
