+++ submissions/TripCraftAI/client/components/footer.tsx
