+++ submissions/TripCraftAI/client/components/header.tsx
